Update how Graphic.Render() works
Previously, Graphic.Render() works in an additive way, any args supplied
(such as rotation, color, flip and so on) will add to Graphic defined
ones before rendering.
But, with this commit, things change to args overriden it's Graphic
counterparts (but it still can be accessed, and used, if needed).
The main reason is be able to provide custom handling to Graphic parameters
with it said, and addition of Transform.Scale (described below), it can affects directly
Graphic.Position and Graphic.Scale at Entity.Render(), without that modification,
a hacky way will be needed

* Add Transform.Scale (it should only affects Transform local space)